# backend/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// BadRequestError Ответ при ошибке 400 Bad Request.
type BadRequestError struct {
	// Code Опциональный машиночитаемый код ошибки.
	Code *string `json:"code,omitempty"`

	// Message Описание причины отклонения запроса.
	Message string `json:"message"`
}

// Booking Подтверждённое бронирование, созданное гостем.
type Booking struct {
	// CreatedAt Временная метка создания записи о бронировании в UTC.
	CreatedAt time.Time `json:"createdAt"`

	// EndTime Время окончания бронирования в UTC.
	EndTime time.Time `json:"endTime"`

	// EventTypeId Идентификатор забронированного типа события.
	EventTypeId Slug `json:"eventTypeId"`

	// EventTypeName Название типа события, сохранённое на момент создания бронирования.
	EventTypeName string `json:"eventTypeName"`

	// GuestEmail Адрес электронной почты гостя.
	GuestEmail string `json:"guestEmail"`

	// GuestName Полное имя гостя.
	GuestName string `json:"guestName"`

	// Id Уникальный идентификатор бронирования, сгенерированный системой (UUID).
	Id string `json:"id"`

	// Note Опциональная заметка или комментарий гостя.
	Note *string `json:"note,omitempty"`

	// StartTime Время начала бронирования в UTC.
	StartTime time.Time `json:"startTime"`
}

// ConflictError Ответ при ошибке 409 Conflict — возвращается, если временной слот уже занят.
type ConflictError struct {
	// ConflictingStartTime Время начала конфликтующего слота в UTC, чтобы гость мог выбрать другой.
	ConflictingStartTime *time.Time `json:"conflictingStartTime,omitempty"`

	// Message Описание конфликта расписания.
	Message string `json:"message"`
}

// CreateBookingRequest Тело запроса для создания бронирования. Гость выбирает тип события и время начала.
type CreateBookingRequest struct {
	// EventTypeId Идентификатор типа события для бронирования.
	EventTypeId Slug `json:"eventTypeId"`

	// GuestEmail Адрес электронной почты гостя.
	GuestEmail string `json:"guestEmail"`

	// GuestName Полное имя гостя.
	GuestName string `json:"guestName"`

	// Note Опциональная заметка или комментарий для владельца.
	Note *string `json:"note,omitempty"`

	// StartTime Желаемое время начала в UTC. Должно соответствовать доступному слоту.
	StartTime time.Time `json:"startTime"`
}

// CreateEventTypeRequest Тело запроса для создания типа события. Владелец явно указывает id (slug).
type CreateEventTypeRequest struct {
	// Description Полное описание события.
	Description string `json:"description"`

	// DurationMinutes Длительность в минутах.
	DurationMinutes int32 `json:"durationMinutes"`

	// Id Slug-идентификатор, выбранный владельцем. Должен быть уникальным. Пример: 'strategy-session'.
	Id Slug `json:"id"`

	// Name Краткое название.
	Name string `json:"name"`
}

// EventType Тип события, созданный владельцем календаря.
type EventType struct {
	// Description Полное описание того, что охватывает событие.
	Description string `json:"description"`

	// DurationMinutes Длительность события в минутах, например 30 или 60.
	DurationMinutes int32 `json:"durationMinutes"`

	// Id Slug-идентификатор, заданный владельцем. Например: 'intro-call'. Уникален среди всех типов событий.
	Id Slug `json:"id"`

	// Name Краткое название, отображаемое гостям.
	Name string `json:"name"`
}

// NotFoundError Ответ при ошибке 404 Not Found.
type NotFoundError struct {
	// Message Описание того, что не было найдено.
	Message string `json:"message"`
}

// Slot Временной слот в окне доступных записей. Возвращаются только свободные слоты.
type Slot struct {
	// EndTime Время окончания слота в UTC, вычисляется как startTime + durationMinutes события.
	EndTime time.Time `json:"endTime"`

	// StartTime Время начала слота в UTC.
	StartTime time.Time `json:"startTime"`
}

// Slug Slug-идентификатор, задаётся владельцем (строчные буквы, цифры, дефисы).
type Slug = string

// GuestApiGetSlotsParams defines parameters for GuestApiGetSlots.
type GuestApiGetSlotsParams struct {
	// Date Дата в формате YYYY-MM-DD.
	Date string `form:"date" json:"date"`
}

// AdminApiCreateEventTypeJSONRequestBody defines body for AdminApiCreateEventType for application/json ContentType.
type AdminApiCreateEventTypeJSONRequestBody = CreateEventTypeRequest

// GuestApiCreateBookingJSONRequestBody defines body for GuestApiCreateBooking for application/json ContentType.
type GuestApiCreateBookingJSONRequestBody = CreateBookingRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /admin/bookings)
	AdminApiListBookings(c *gin.Context)

	// (DELETE /admin/bookings/{id})
	AdminApiDeleteBooking(c *gin.Context, id string)

	// (GET /admin/event-types)
	AdminApiListEventTypes(c *gin.Context)

	// (POST /admin/event-types)
	AdminApiCreateEventType(c *gin.Context)

	// (POST /bookings)
	GuestApiCreateBooking(c *gin.Context)

	// (GET /event-types)
	GuestApiListEventTypes(c *gin.Context)

	// (GET /event-types/{id}/available-days)
	GuestApiGetAvailableDays(c *gin.Context, id Slug)

	// (GET /event-types/{id}/slots)
	GuestApiGetSlots(c *gin.Context, id Slug, params GuestApiGetSlotsParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// AdminApiListBookings operation middleware
func (siw *ServerInterfaceWrapper) AdminApiListBookings(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AdminApiListBookings(c)
}

// AdminApiDeleteBooking operation middleware
func (siw *ServerInterfaceWrapper) AdminApiDeleteBooking(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AdminApiDeleteBooking(c, id)
}

// AdminApiListEventTypes operation middleware
func (siw *ServerInterfaceWrapper) AdminApiListEventTypes(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AdminApiListEventTypes(c)
}

// AdminApiCreateEventType operation middleware
func (siw *ServerInterfaceWrapper) AdminApiCreateEventType(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AdminApiCreateEventType(c)
}

// GuestApiCreateBooking operation middleware
func (siw *ServerInterfaceWrapper) GuestApiCreateBooking(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GuestApiCreateBooking(c)
}

// GuestApiListEventTypes operation middleware
func (siw *ServerInterfaceWrapper) GuestApiListEventTypes(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GuestApiListEventTypes(c)
}

// GuestApiGetAvailableDays operation middleware
func (siw *ServerInterfaceWrapper) GuestApiGetAvailableDays(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id Slug

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GuestApiGetAvailableDays(c, id)
}

// GuestApiGetSlots operation middleware
func (siw *ServerInterfaceWrapper) GuestApiGetSlots(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id Slug

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GuestApiGetSlotsParams

	// ------------- Required query parameter "date" -------------

	if paramValue := c.Query("date"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument date is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", false, true, "date", c.Request.URL.Query(), &params.Date, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter date: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GuestApiGetSlots(c, id, params)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/admin/bookings", wrapper.AdminApiListBookings)
	router.DELETE(options.BaseURL+"/admin/bookings/:id", wrapper.AdminApiDeleteBooking)
	router.GET(options.BaseURL+"/admin/event-types", wrapper.AdminApiListEventTypes)
	router.POST(options.BaseURL+"/admin/event-types", wrapper.AdminApiCreateEventType)
	router.POST(options.BaseURL+"/bookings", wrapper.GuestApiCreateBooking)
	router.GET(options.BaseURL+"/event-types", wrapper.GuestApiListEventTypes)
	router.GET(options.BaseURL+"/event-types/:id/available-days", wrapper.GuestApiGetAvailableDays)
	router.GET(options.BaseURL+"/event-types/:id/slots", wrapper.GuestApiGetSlots)
}