# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Calendar Booking API
  description: 'REST API for a calendar booking application. Supports two roles: the calendar owner (admin) and anonymous guests.'
  version: 0.0.0
tags: []
paths:
  /admin/bookings:
    get:
      operationId: AdminApi_listBookings
      description: |2-
          Получить список предстоящих бронирований по всем типам событий, отсортированных по startTime по возрастанию.
          Предстоящие — те, у которых startTime не раньше текущего времени сервера.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
  /admin/bookings/{id}:
    delete:
      operationId: AdminApi_deleteBooking
      description: Удалить бронирование по id. Возвращает 404, если бронирование не найдено.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /admin/event-types:
    post:
      operationId: AdminApi_createEventType
      description: |2-
          Создать новый тип события.
          Владелец указывает id (slug) в теле запроса.
          Возвращает 400, если id уже занят или входные данные некорректны.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventTypeRequest'
    get:
      operationId: AdminApi_listEventTypes
      description: Получить список всех типов событий, созданных владельцем.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
  /bookings:
    post:
      operationId: GuestApi_createBooking
      description: |2-
          Создать бронирование на выбранный слот.
          Проверяемые бизнес-правила:
          - Слот должен приходиться на будний день (Пн–Пт) в диапазоне 09:00–18:00.
          - Слот должен попадать в окно бронирования — 14 дней от сегодня.
          - Ни одно другое бронирование (любого типа события) не должно перекрывать запрошенный диапазон.
          Возвращает 400 при некорректных входных данных, 404 если тип события не найден,
          и 409 Conflict если слот уже занят.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
  /event-types:
    get:
      operationId: GuestApi_listEventTypes
      description: 'Получить список всех типов событий в публичном виде: id, name, description, durationMinutes.'
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
  /event-types/{id}/available-days:
    get:
      operationId: GuestApi_getAvailableDays
      description: |2-
          Получить список дат (формат YYYY-MM-DD) на ближайшие 14 дней, в которые есть хотя бы один свободный слот.
          Используется для первичной отрисовки календаря. Возвращает 404, если тип события не найден.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Slug'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /event-types/{id}/slots:
    get:
      operationId: GuestApi_getSlots
      description: |2-
          Получить свободные слоты для указанного типа события на конкретную дату (YYYY-MM-DD).
          Слоты генерируются по будням с 09:00 до 18:00 (локальное время сервера).
          Возвращаются только слоты без конфликтующих бронирований.
          Возвращает 404, если тип события не найден.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Slug'
        - name: date
          in: query
          required: true
          description: Дата в формате YYYY-MM-DD.
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
components:
  schemas:
    BadRequestError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Описание причины отклонения запроса.
        code:
          type: string
          description: Опциональный машиночитаемый код ошибки.
      description: Ответ при ошибке 400 Bad Request.
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - eventTypeName
        - guestName
        - guestEmail
        - startTime
        - endTime
        - createdAt
      properties:
        id:
          type: string
          description: Уникальный идентификатор бронирования, сгенерированный системой (UUID).
        eventTypeId:
          allOf:
            - $ref: '#/components/schemas/Slug'
          description: Идентификатор забронированного типа события.
        eventTypeName:
          type: string
          description: Название типа события, сохранённое на момент создания бронирования.
        guestName:
          type: string
          description: Полное имя гостя.
        guestEmail:
          type: string
          description: Адрес электронной почты гостя.
        note:
          type: string
          description: Опциональная заметка или комментарий гостя.
        startTime:
          type: string
          format: date-time
          description: Время начала бронирования в UTC.
        endTime:
          type: string
          format: date-time
          description: Время окончания бронирования в UTC.
        createdAt:
          type: string
          format: date-time
          description: Временная метка создания записи о бронировании в UTC.
      description: Подтверждённое бронирование, созданное гостем.
    ConflictError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Описание конфликта расписания.
        conflictingStartTime:
          type: string
          format: date-time
          description: Время начала конфликтующего слота в UTC, чтобы гость мог выбрать другой.
      description: Ответ при ошибке 409 Conflict — возвращается, если временной слот уже занят.
    CreateBookingRequest:
      type: object
      required:
        - eventTypeId
        - startTime
        - guestName
        - guestEmail
      properties:
        eventTypeId:
          allOf:
            - $ref: '#/components/schemas/Slug'
          description: Идентификатор типа события для бронирования.
        startTime:
          type: string
          format: date-time
          description: Желаемое время начала в UTC. Должно соответствовать доступному слоту.
        guestName:
          type: string
          description: Полное имя гостя.
        guestEmail:
          type: string
          description: Адрес электронной почты гостя.
        note:
          type: string
          description: Опциональная заметка или комментарий для владельца.
      description: Тело запроса для создания бронирования. Гость выбирает тип события и время начала.
    CreateEventTypeRequest:
      type: object
      required:
        - id
        - name
        - description
        - durationMinutes
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Slug'
          description: "Slug-идентификатор, выбранный владельцем. Должен быть уникальным. Пример: 'strategy-session'."
        name:
          type: string
          description: Краткое название.
        description:
          type: string
          description: Полное описание события.
        durationMinutes:
          type: integer
          format: int32
          description: Длительность в минутах.
      description: Тело запроса для создания типа события. Владелец явно указывает id (slug).
    EventType:
      type: object
      required:
        - id
        - name
        - description
        - durationMinutes
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Slug'
          description: "Slug-идентификатор, заданный владельцем. Например: 'intro-call'. Уникален среди всех типов событий."
        name:
          type: string
          description: Краткое название, отображаемое гостям.
        description:
          type: string
          description: Полное описание того, что охватывает событие.
        durationMinutes:
          type: integer
          format: int32
          description: Длительность события в минутах, например 30 или 60.
      description: Тип события, созданный владельцем календаря.
    NotFoundError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Описание того, что не было найдено.
      description: Ответ при ошибке 404 Not Found.
    Slot:
      type: object
      required:
        - startTime
        - endTime
      properties:
        startTime:
          type: string
          format: date-time
          description: Время начала слота в UTC.
        endTime:
          type: string
          format: date-time
          description: Время окончания слота в UTC, вычисляется как startTime + durationMinutes события.
      description: Временной слот в окне доступных записей. Возвращаются только свободные слоты.
    Slug:
      type: string
      description: Slug-идентификатор, задаётся владельцем (строчные буквы, цифры, дефисы).
servers:
  - url: http://localhost:8080
    description: Local development server
    variables: {}